refactor(db): move usage store ownership
